requirements.txt
torch
torchvision
facenet-pytorch
transformers
tqdm
numpy
av
librosa
opencv-python
scikit-learn
openai-whisper